requirements.txt
boto
paramiko
mysql-connector
loguru
prettytable